divide contracts.go to multiple files
